+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # INT TEAM API CHANGELOG
 This changelog only contains overviews. For further reading, always consult the current wiki/documentation.
+## v5.0.1 (7th Nov 2023)
+### IntAPI:
+- OnFirstSpawn class is now private. Use **IntAPI.IsFirstSpawn(GameObject gameObject)** instead.
+### LimbPlus:
+- Added gorse limbs to LimbTypes
+- Made FindLimb() & FindLimbComp() more concise, only 1 of these overloads has the actual behaviour
 ## v5.0.0 (5th Nov 2023)
 ### General:
 - Since this is technically a new release, no changelog will be provided, and thank god because if I had to actually figure out what changed and got added, I would lose my sanity.